+++ kernel/src/Data-structures/AVLTree.hpp
@@ -10,7 +10,7 @@ namespace AVLTree {
 	struct Node {
 		// actual data
 		uint64_t key;
-		void* extraData;
+		uint64_t extraData;
 
 		// Required extras
 		Node* left;
@@ -27,7 +27,7 @@ namespace AVLTree {
 	uint64_t height(Node* root);
 
 	// Helper function that allocates a new node with the given key, extra data and NULL left and right pointers.
-	Node* newNode(uint64_t key, void* extraData);
+	Node* newNode(uint64_t key, uint64_t extraData);
 
 	// A utility function to right rotate subtree.
 	Node* rightRotate(Node* root);
@@ -39,10 +39,13 @@ namespace AVLTree {
 	int64_t getBalance(Node* N);
 
 	// Recursive function to insert a key in the subtree and returns the new root of the subtree.
-	void insert(Node*& root, uint64_t key, void* extraData);
+	void insert(Node*& root, uint64_t key, uint64_t extraData);
 
-	// Get a pointer a node from its key
+	// Get a pointer to a node from its key
 	Node* findNode(Node* root, uint64_t key);
+
+	// Get a pointer to a node from its key or the key with the next highest value
+	Node* findNodeOrHigher(Node* root, uint64_t key);
 
 	// Get the node with the lowest value in the tree
 	Node* minValueNode(Node* root);
@@ -53,6 +56,9 @@ namespace AVLTree {
 
 	// A utility function to print preorder traversal of the tree.
 	void preOrder(Node* root);
+
+	// find Parent node of a key
+	Node* getParent(Node* root, uint64_t key);
 
 }