+++ kernel/src/Memory/newdelete.cpp
@@ -48,6 +48,10 @@ void operator delete[](void* p) {
     }
 }
 
+void operator delete(void* p, size_t) {
+    delete p;
+}
+
 
 bool g_NewDeleteInitialised = false;