+++ kernel/src/Memory/newdelete.hpp
@@ -5,7 +5,7 @@
 
 // standard new/delete functions
 
-void* operator new(size_t size);
+void* operator new(size_t size) throw();
 void* operator new[](size_t size);
 
 void operator delete(void* p);
@@ -14,7 +14,7 @@ void operator delete[](void* p);
 
 // following functions aren't part of the C++ ABI, but are required to fix linker errors
 
-inline void operator delete(void* p, size_t) { delete p; };
+void operator delete(void* p, size_t);
 
 
 // utility functions