+++ kernel/src/arch/x86_64/Stack.c
@@ -1,0 +1,3 @@
+#include "Stack.h"
+
+unsigned char kernel_stack[65536] = {0};